test: bool upcast when comparing with integers + updating OOB tests to use literals directly

+++ test_cases/regression_tests/getbit_fold_oob_uint64.py
@@ -1,10 +1,10 @@
-from algopy import Contract, UInt64, op
+from algopy import Contract, op
 
 
 class GetBitFoldOOBUInt64(Contract):
     def approval_program(self) -> bool:
         # out of bounds bit index access (should be indices 0...63)
-        assert op.getbit(0, 64) == UInt64(0)
+        assert op.getbit(0, 64) == 0
         return True
 
     def clear_state_program(self) -> bool:
